+++ app/src/main/java/org/android/go/sopt/data/model/FakeGithubInfo.kt
@@ -1,3 +1,3 @@
-package org.android.go.sopt.model
+package org.android.go.sopt.data.model
 
 class FakeGithubInfo : ArrayList<FakeGithubInfoItem>()
