+++ app/src/main/java/org/android/go/sopt/data/model/FakeGithubInfoItem.kt
@@ -1,4 +1,4 @@
-package org.android.go.sopt.model
+package org.android.go.sopt.data.model
 
 import com.google.gson.annotations.SerializedName
