+++ app/src/main/java/org/android/go/sopt/data/model/Owner.kt
@@ -1,4 +1,4 @@
-package org.android.go.sopt.model
+package org.android.go.sopt.data.model
 
 data class Owner(
     val avatar_url: String,
