+++ app/src/main/java/org/android/go/sopt/data/model/UserInfo.kt
@@ -1,4 +1,4 @@
-package org.android.go.sopt.model
+package org.android.go.sopt.data.model
 
 data class UserInfo(
     val id: String,
