+++ app/src/main/java/org/android/go/sopt/di/RepoModule.kt
@@ -4,16 +4,21 @@ import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
+import org.android.go.sopt.data.repoImpl.HomeRepoImpl
 import org.android.go.sopt.domain.SignRepository
-import org.android.go.sopt.model.repoImpl.SignRepoImpl
+import org.android.go.sopt.data.repoImpl.SignRepoImpl
+import org.android.go.sopt.domain.HomeRepository
 import javax.inject.Singleton
 
 @Module
 @InstallIn(SingletonComponent::class)
 abstract class RepositoryModule {
-
     @Singleton
     @Binds
     abstract fun providesSignRepo(repoImpl: SignRepoImpl): SignRepository
+
+    @Singleton
+    @Binds
+    abstract fun providesHomeRepo(repoImpl: HomeRepoImpl): HomeRepository
 
 }
