+++ app/src/main/java/org/android/go/sopt/ui/join/JoinActivity.kt
@@ -9,8 +9,8 @@ import dagger.hilt.android.AndroidEntryPoint
 import org.android.go.sopt.App
 import org.android.go.sopt.R
 import org.android.go.sopt.databinding.ActivityJoinBinding
-import org.android.go.sopt.model.RequestSignUpDto
-import org.android.go.sopt.model.UserInfo
+import org.android.go.sopt.data.model.RequestSignUpDto
+import org.android.go.sopt.data.model.UserInfo
 import org.android.go.sopt.util.Constants.INPUT_SUCCESS
 import org.android.go.sopt.util.afterTextChanged
 import org.android.go.sopt.util.hideKeyboard
@@ -68,11 +68,13 @@ class JoinActivity : AppCompatActivity() {
     private fun clickJoin() {
         setUser()
         binding.btnJoin.setOnClickListener {
-            viewModel.signUp(RequestSignUpDto(
+            viewModel.signUp(
+                RequestSignUpDto(
                 binding.etvId.text.toString(),
                 binding.etvPwd.text.toString(),
                 binding.etvName.text.toString(),
-                binding.etvSpecialty.text.toString()))
+                binding.etvSpecialty.text.toString())
+            )
             val intent = Intent().apply {
                 putExtra("id", binding.etvId.text.toString())
                 putExtra("password", binding.etvPwdCheck.text.toString())
